Use pre-parsed git options
'git CMD -h' is no longer used to get the list of short and long
options for completion because it is usually incomplete and misses a
few key commands such as rebase diff log etc.

Git command options are now pre-parsed by an elisp script and saved to
a file designated by pcmpl-git-options-file. This offers a better
coverage and accuracy.

Git config variables are also parsed and saved.